Create Credit
This code lets you detect a fake credit card by checking the credit card number to see if it follows the credit card number rules, if it doesn't, the program rejects the card and detects a wrong credit card number. This code can be used in banks, stores, restaurants, airports, basically everywhere you need a credit card payment. It can be used to catch a criminal or save an organization from losses.